refactor(exoplayer): extract timer wrapper
